+++ src/components/main/main.e2e.test.js
@@ -5,7 +5,6 @@ import {Provider} from 'react-redux';
 import configureStore from 'redux-mock-store';
 
 import Main from "./main.jsx";
-import {PromoData} from "../../consts.js";
 
 const mockStore = configureStore([]);
 
@@ -20,6 +19,18 @@ const mockMovies = [
   {title: `Macbeth`, poster: `http://placekitten.com/245/175`},
 ];
 
+const movie = {title: `Mindhunter`, poster: `http://placekitten.com/245/175`, cover: `img/bg-the-grand-budapest-hotel.jpg`, genre: `drama`, releaseYear: 2000, src: ``, rating: 5, ratingNumber: 100, ratingValue: ``, starring: [``, `1`, `3`], director: ``, reviews: [{author: `Amanda Greever`,
+  date: `November 18, 2015`,
+  id: 1,
+  message: `The mannered, madcap proceedings are often delightful, occasionally silly, and here and there, gruesome and/or heartbreaking.`,
+  rating: 2},
+{author: `Amanda Greever`,
+  date: `November 18, 2015`,
+  id: 2,
+  message: `The mannered, madcap proceedings are often delightful, occasionally silly, and here and there, gruesome and/or heartbreaking.`,
+  rating: 2}]};
+
+
 const mockGenres = [`All genres`, `Drama`, `Comedy`, `Thriller`, `Fantasy`, `Horror`];
 
 Enzyme.configure({
@@ -32,15 +43,14 @@ it(`Movie card title clicked`, () => {
     movies: mockMovies,
     genresList: mockGenres,
     shownMoviesNumber: 8,
+    promoMovie: movie,
   });
   const onTitleClick = jest.fn();
 
   const mainPage = shallow(
       <Provider store={store}>
         <Main
-          movieName = {PromoData.movieName}
-          movieGenre = {PromoData.movieGenre}
-          movieReleaseDate = {PromoData.movieReleaseDate}
+          promoMovie={movie}
           movies = {mockMovies}
           onTitleClick = {onTitleClick}
           genresList = {mockGenres}
